Break smallsh down further into organized modules
