+++ utilities.c
@@ -1,0 +1,69 @@
+/**
+ * Definitions for various utility functions.
+*/
+
+#include <err.h>
+#include <errno.h>
+#include <stdio.h>
+#include <stdlib.h>
+#include <string.h>
+#include <unistd.h>
+
+#include "utilities.h"
+
+/**
+ * https://edstem.org/us/courses/21025/discussion/1437434?answer=3250618
+ * Converts the current process ID into a string and returns the pointer
+ * to it.
+ *
+ * @return The pointer to the PID string
+ */
+char *
+get_pidstr(void)
+{
+  pid_t pid = getpid();
+  char *pidstr;
+  {
+    int n = snprintf(NULL, 0, "%d", pid);
+    pidstr = malloc((n + 1) * sizeof *pidstr);
+    sprintf(pidstr, "%d", pid);
+  }
+  return pidstr;
+}
+
+/**
+ * (from archive.c skeleton code)
+ * Allocates a string containing the CWD
+ *
+ * @return allocated string
+ */
+char *
+getcwd_a(void)
+{
+  char *pwd = NULL;
+  for (size_t sz = 128;; sz *= 2)
+  {
+    pwd = realloc(pwd, sz);
+    if (getcwd(pwd, sz)) break;
+    if (errno != ERANGE) err(errno, "getcwd()");
+  }
+
+  return pwd;
+}
+
+/**
+ * https://edstem.org/us/courses/21025/discussion/1360952?answer=3085214
+ * Duplicates the given string and returns a pointer to the copy.
+ *
+ * @param s The pointer to the string to duplicate
+ * @return The pointer to the copy of the string
+ */
+char *
+strdup (const char *s)
+{
+  size_t len = strlen (s) + 1;
+  void *new = malloc (len);
+  if (new == NULL)
+    return NULL;
+  return (char *) memcpy (new, s, len);
+}
