+++ utilities.h
@@ -1,0 +1,8 @@
+#ifndef UTILITIES_H
+#define UTILITIES_H
+
+char * get_pidstr(void);
+char * getcwd_a(void);
+char * strdup (const char *s);
+
+#endif
